requirements.txt
numpy==1.23.1
opencv_python_headless==4.6.0.66
Pillow==9.4.0
requests==2.25.1
streamlit==1.12.2
streamlit_lottie==0.0.3
streamlit_option_menu==0.3.2
streamlit_webrtc==0.44.2
tensorflow==2.10.0